app/requirements.txt
fastapi
uvicorn
python-arango
pydantic